対局中のViewを分離
Previewでレイアウト調整するため
